+++ CHANGELOG.md
@@ -5,5 +5,6 @@
 Released on .
 
 - Fix `--filter-entry-types` option in `show` command.
+- Fix Path.walk() not working in versions of python <= 3.11. Now it uses os.walk() if Path.walk() not available.
 
 Pypi: https://pypi.org/project/bibmancli/0.3.4/